requirements.txt
pandas
regex
dash==1.13.4
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
Flask==1.1.2
Flask-Compress==1.5.0
Jinja2==2.11.2
keyring==21.2.1
keyrings.alt
mysql-connector-python==8.0.28
plotly==4.8.2
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pywin32-ctypes==0.2.0
regex==2020.6.8
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
tqdm==4.47.0
itsdangerous==1.1.0
MarkupSafe==1.1.1
Werkzeug==1.0.1
yagmail==0.11.224
